Move read_path_if_provided to utils
