+++ src/utils.py
@@ -1,7 +1,50 @@
 import base64
 
+import os
 import pandas as pd
 import plotly.express as px
+
+def read_path_if_provided(spark, path):
+    """Automatically detect the format of the input data and read it into the Spark dataframe. The supported formats
+    are: a single TSV file; a single JSON file; a directory with JSON files; a directory with Parquet files."""
+    # All datasets are optional.
+    if path is None:
+        return None
+
+    # The provided path must exist and must be either a file or a directory.
+    assert os.path.exists(path), f'The provided path {path} does not exist.'
+    assert os.path.isdir(path) or os.path.isfile(path), \
+        f'The provided path {path} is neither a file or a directory.'
+
+    # Case 1: We are provided with a single file.
+    if os.path.isfile(path):
+        if path.endswith('.tsv'):
+            return spark.read.csv(path, sep='\t', header=True)
+        elif path.endswith(('.json', '.json.gz', '.jsonl', '.jsonl.gz')):
+            return spark.read.json(path)
+        else:
+            raise AssertionError(f'The format of the provided file {path} is not supported.')
+
+    # Case 2: We are provided with a directory. Let's peek inside to see what it contains.
+    all_files = [
+        os.path.join(dp, filename)
+        for dp, dn, filenames in os.walk(path)
+        for filename in filenames
+    ]
+
+    # It must be either exclusively JSON, or exclusively Parquet.
+    json_files = [fn for fn in all_files if fn.endswith(('.json', '.json.gz', '.jsonl', '.jsonl.gz'))]
+    parquet_files = [fn for fn in all_files if fn.endswith('.parquet')]
+    assert not(json_files and parquet_files), f'The provided directory {path} contains a mix of JSON and Parquet.'
+    assert json_files or parquet_files, f'The provided directory {path} contains neither JSON nor Parquet.'
+
+    # A directory with JSON files.
+    if json_files:
+        return spark.read.option('recursiveFileLookup', 'true').json(path)
+
+    # A directory with Parquet files.
+    if parquet_files:
+        return spark.read.parquet(path)
 
 def get_table_download_link_csv(df):
     csv = df.to_csv().encode()
